Created ojph_threads. Rename socket to sockets, and fixed a warning.
